Refactor function to color file.
